requirements.txt
autopep8==1.6.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
customtkinter==4.1.0
darkdetect==0.6.0
et-xmlfile==1.1.0
gTTS==2.2.4
idna==3.3
jdcal==1.4.1
mysql-connector-python==8.0.29
numpy==1.18.0
opencv-contrib-python==4.5.5.64
opencv-python==4.5.5.64
openpyxl==3.0.3
pandas==1.0.3
Pillow==7.1.1
playsound==1.3.0
protobuf==4.21.1
pycodestyle==2.8.0
pypiwin32==223
python-dateutil==2.8.2
pyttsx3==2.71
pytz==2022.1
pywin32==304
requests==2.27.1
six==1.16.0
toml==0.10.2
urllib3==1.26.9
